refactor, add feature products
