+++ src/Components/FeaturedProducts.js
@@ -1,70 +1,28 @@
-import React, { Component, Grid, Row, Col, Thumbnail, Button } from 'react'
+import React, { Component } from 'react'
+import FeaturedProductsItem from './FeaturedProductsItem'
+import { Row, Col } from 'react-bootstrap'
 
-import getWeb3 from '../utils/getWeb3'
-import { getPublicMarketContract } from '../utils/contracts'
 
 class FeaturedProducts extends Component {
+  constructor(props) {
+    super(props)
+  }
 
   render() {
+    var items = []
+    this.props.dins.forEach(function(din, index) {
+      items.push(<FeaturedProductsItem key={index} din={din} />)
+    })
+
     return (
-      <Grid>
-        <Row>
-          {
-            this.props.dins.map(function(din) {
-              return <FeaturedProductsItem din={din} />
-            })
-          }
-        </Row>
-      </Grid>
+      <div>
+        <Col md={4} mdOffset={4}>
+          <h2 className="text-center"> Featured Products </h2>
+          {items}
+        </Col>
+      </div>
     )
   }
 }
 
-class FeaturedProductsItem extends Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      web3: null,
-      price: null,
-      formattedPrice: null,
-      name: null
-    }
-  }
-
-  componentWillMount() {
-    getWeb3.then(results => {
-      this.setState({
-        web3: results.web3
-      })
-
-      this.initializeResolver()
-    })
-  }
-
-  initializeResolver() {
-    getPublicMarketContract.then(contract => {
-      this.setState({ publicMarket: contract })
-      contract.price(this.props.din).then((price) => {
-        this.setState({ price: price.toNumber() })
-        let formattedPrice = this.state.web3.fromWei(price, 'ether') + " ether"
-        this.setState({ formattedPrice: formattedPrice })
-      })
-
-
-    })
-  }
-
-  render() {
-    <Col xs={6} md={4}>
-      <Thumbnail src="/assets/thumbnaildiv.png" alt="242x200">
-        <h3>Thumbnail label</h3>
-        <p>Description</p>
-        <p><Button bsStyle="primary">Buy</Button></p>
-      </Thumbnail>
-    </Col>
-  }
-}
-
-// export { FeaturedProducts, FeaturedProductsItem }
-
-export default FeaturedProducts;
+export default FeaturedProducts
